Users/api/serializers/user_serializer.py
from rest_framework.serializers import serializers
from Users.models import CustomUser

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = ['username', 'email']

