requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
django-cors-headers==3.10.0
django==3.2.8
djangorestframework==3.12.4
pytz==2021.3
sqlparse==0.4.2; python_version >= '3.5'
